pyproject.toml
[tool.poetry]
name = "pennapps-backend"
version = "0.1.0"
description = ""
authors = ["Punit Arani <punitsai36@gmail.com>"]
readme = "README.md"
packages = [{include = "pennapps_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
docker = "^6.1.3"
openai = "^0.28.0"
pandas = "^2.1.0"
pyarrow = "^13.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
scikit-learn = "^1.3.0"
supabase = "^1.0.4"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
